refactor(devtools): 重构开发者工具路径查找逻辑
