+++ pnpm-workspace.yaml
@@ -14,6 +14,8 @@ catalog:
   kolorist: ^1.8.0
   local-pkg: ^1.1.2
   moment: ^2.30.1
+  pathe: ^2.0.3
+  std-env: ^3.10.0
   unconfig: ^7.3.3
 catalogs:
   cli:
