+++ src/constants/index.ts
@@ -1,2 +1,3 @@
+export * from './devtools'
 export * from './platform'
 export * from './terminal'
